requirements.txt
ultralytics
tkinter
PIL
pyyaml